웹 메인 화면과 대쉬보드 분리 등 파일 구조 및 라우터 수정/추가
