+++ Platforms/NodeJS-MQTT/views/auth/login.ejs
@@ -15,9 +15,8 @@
 
 <a href="/auth/facebook">facebook 로그인</a>
 <a href="/auth/kakao">kakao 로그인</a>
+<a href="/auth/register">회원가입</a>
 
 <% if(typeof err != 'undefined') { %>
-    Login error = <%=err%>
+Login error = <%=err%>
 <%}%>
-
-<a href="/auth/register">회원가입</a>
