+++ Platforms/NodeJS-MQTT/views/dashboard/board.ejs
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <% include ../header.ejs %>
+    </head>
+    <body>
+        <% include ./header.ejs%>
+
+        <div class="menu">
+            <a href="/logging">메세지 로그</a>
+            <a href="/board">메세지 발행</a>
+            <a href="/nodeinfo">노드 관리</a>
+        </div>
+        <div class="content">
+            <% if(page === 'logging') { %>
+            <% include ./logging.ejs %>
+            <% } else { %>
+            <% include ./publish.ejs %>
+            <% }%>
+        </div>
+    </body>
+</html>
