+++ Platforms/NodeJS-MQTT/views/dashboard/header.ejs
@@ -1,0 +1,6 @@
+<div>
+    <h1>Board</h1>
+    <% %>
+    <br>
+    <a href="/auth/logout">Logout</a>
+</div>
