+++ Platforms/NodeJS-MQTT/views/dashboard/publish.ejs
